Fixed broken link to Rob Sewell's blog
Link to Rob Sewell's blog article called "How to fork a GitHub repository and contribute to an open source project" was broken. I updated the link to the correct location and added a mirror link to the article on archive.org, in case the URL changes again in the future.